Add new option to load files to PCI device ('m').
The current 'f' option is a special case for loading an eBPF program and
should not be used in general (it does not write data exactly where you
specify in the command line and also writes the binary file length).

The new option, 'm' (for memory), should be used instead, as it writes
the whole file in the location specified.

Signed-off-by: Martin Ichilevici de Oliveira <martin.i.oliveira@gmail.com>